Refactor experiments and alerts pages
