+++ frontend/components/ui/jobs/JobCard.js
@@ -1,33 +1,16 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Head from 'next/head'
-import Link from 'next/link'
-
-import Immutable from 'immutable'
-
-import Avatar from 'material-ui/Avatar'
-
-import Button from 'material-ui/Button'
-import AddIcon from '@material-ui/icons/Add'
-import ClearIcon from '@material-ui/icons/Clear'
 import DeleteIcon from '@material-ui/icons/Delete'
 import MessageIcon from '@material-ui/icons/Message'
 import AssignmentIcon from '@material-ui/icons/Assignment'
+
 import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp'
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown'
 
 import Card, { CardHeader, CardContent, CardActions } from 'material-ui/Card'
 
 import List, { ListItem, ListItemText } from 'material-ui/List'
-
-import moment from 'moment'
-
-import { Flex, Box } from 'ooni-components'
-import { Container } from 'ooni-components'
-
-import Layout from '../../../components/layout'
-import Session from '../../../components/session'
 
 class JobCard extends React.Component {
   static propTypes = {
@@ -166,107 +149,4 @@ class JobCard extends React.Component {
   }
 }
 
-class AdminJobsIndex extends React.Component {
-  constructor (props) {
-    super(props)
-    this.state = {
-      jobList: [],
-      error: null,
-      session: new Session(),
-      actionButtonOpen: false
-    }
-    this.onDelete = this.onDelete.bind(this)
-    this.toggleAction = this.toggleAction.bind(this)
-  }
-
-  toggleAction () {
-    this.setState({actionButtonOpen: !this.state.actionButtonOpen})
-  }
-
-  onDelete (jobId) {
-    let req = this.state.session.createRequest({baseURL: process.env.ORCHESTRATE_URL})
-    req.delete(`/api/v1/admin/job/${jobId}`)
-      .then((res) => {
-        const newJobList = this.state.jobList.filter((job) => (job.id !== jobId))
-        this.setState({
-          jobList: newJobList
-        })
-      })
-      .catch((err) => {
-        // XXX handle errors
-      })
-  }
-
-  componentDidMount() {
-    if (this.state.session.redirectIfInvalid()) {
-      return
-    }
-    let req = this.state.session.createRequest({baseURL: process.env.ORCHESTRATE_URL})
-    req.get('/api/v1/admin/jobs')
-      .then((res) => {
-        this.setState({
-          jobList: res.data.jobs || [],
-          error: null
-        })
-    }).catch((err) => {
-        this.setState({
-          error: err
-        })
-    })
-  }
-
-  render () {
-    const {
-      jobList,
-      actionButtonOpen
-    } = this.state
-
-    return (
-      <Layout title="Jobs">
-        <Head>
-          <title>Jobs - OONI Proteus</title>
-        </Head>
-
-        <div>
-          <Container style={{position: 'relative'}}>
-            {jobList.map((job) => {
-              return (
-                <Flex>
-                <JobCard
-                  key={job.id}
-                  onDelete={this.onDelete}
-                  comment={job.comment}
-                  creationTime={job.creation_time}
-                  delay={job.delay}
-                  id={job.id}
-                  schedule={job.schedule}
-                  state={job.state}
-                  target={job.target}
-                  alertData={job.alert}
-                  task={job.task} />
-                </Flex>
-              )
-            })}
-            <div>
-              <Flex column>
-                {actionButtonOpen && <Box pt={2}>
-                  <Link href='/admin/jobs/add_alert'><Button variant="fab" mini><MessageIcon/></Button></Link>
-                </Box>}
-                {actionButtonOpen && <Box pt={2}>
-                  <Link href='/admin/jobs/add_task'><Button variant="fab" mini><AssignmentIcon/></Button></Link>
-                </Box>}
-
-                <Box pt={2} onClick={() => this.toggleAction()}>
-                  {!actionButtonOpen && <Button variant="fab" color="primary" mini><AddIcon/></Button>}
-                  {actionButtonOpen && <Button variant="fab" color="primary" mini><ClearIcon/></Button>}
-                </Box>
-              </Flex>
-            </div>
-          </Container>
-        </div>
-      </Layout>
-    )
-  }
-}
-
-export default AdminJobsIndex
+export default JobCard
